Customized Custom Vector Type
- Axes are public, can be accessed easier for e.g. .x, .y
- Able to support addition and subtraction operations with glm vec (will return eclipse vector itself)
- Able to convert eclipse vector into glmvecs of 2,3 and 4 by function calls

+++ Eclipse/src/Graphics/InputHandler/InputWrapper.cpp
@@ -393,7 +393,7 @@ bool InputWrapper::GetKeyTriggered(InputKeycode keycode)
 
       if (EnablePrint)
       {
-        std::cout << "Pressed " << (Message) << " " << std::endl;
+        std::cout << "Pressed " << static_cast<int>(Message) << " " << std::endl;
       }
       return true;
     }
@@ -462,7 +462,7 @@ bool Eclipse::InputWrapper::GetKeyReleased(InputKeycode keycode)
 
         if (EnablePrint)
         {
-          std::cout << "Released " << (Message) << " " << std::endl;
+          std::cout << "Released " << static_cast<int>(Message) << " " << std::endl;
         }
         return true;
       }
